Project Structure/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2
pydantic
Pillow
python-multipart
python-jose
passlib
bcrypt==4.0.1
